+++ Backend-filmhub/app/Http/Controllers/admin/FoodController.php
@@ -64,6 +64,9 @@ class FoodController extends Controller
     public function destroy($id)
     {
         $food = Food::findOrFail($id);
+        if ($food->tickets()->exists()) {
+            return redirect()->route('admin.foods.index')->with('error', 'Không thể xóa món ăn vì đã có vé liên quan.');
+        }
         $food->delete();
         return redirect()->route('admin.foods.index')->with('success', 'Food deleted successfully.');
     }
